[svn-r11307] Purpose:     Code cleanup
Description:
    Clean up internals of group creation & iteration code.

Platforms tested:
    FreeBSD 4.11 (sleipnir)
    Mac OS X (nile)
    Too minor to require h5committest

+++ src/H5Fprivate.h
@@ -52,13 +52,13 @@ typedef struct H5F_t H5F_t;
  */
 
 #  define INT16ENCODE(p, i) {						      \
-   *(p) = (uint8_t)( (unsigned)(i)	    & 0xff); (p)++;			      \
-   *(p) = (uint8_t)(((unsigned)(i) >> 8) & 0xff); (p)++;			      \
+   *(p) = (uint8_t)( (unsigned)(i)	 & 0xff); (p)++;		      \
+   *(p) = (uint8_t)(((unsigned)(i) >> 8) & 0xff); (p)++;		      \
 }
 
 #  define UINT16ENCODE(p, i) {						      \
-   *(p) = (uint8_t)(	    (i)	    & 0xff); (p)++;			      \
-   *(p) = (uint8_t)(((unsigned)(i) >> 8) & 0xff); (p)++;			      \
+   *(p) = (uint8_t)( (unsigned)(i)	 & 0xff); (p)++;		      \
+   *(p) = (uint8_t)(((unsigned)(i) >> 8) & 0xff); (p)++;		      \
 }
 
 #  define INT32ENCODE(p, i) {						      \
